Refactor Node and Filters classes to use proper type names
