+++ src/index.ts
@@ -3,3 +3,9 @@ export * from "./structures/Node";
 export * from "./structures/Player";
 export * from "./structures/Queue";
 export * from "./structures/Utils";
+export * from "./structures/Filters";
+export * from "./structures/Rest";
+export * from "./utils/FiltersEqualizers";
+export * from "./utils/ManagerCheck";
+export * from "./utils/NodeCheck";
+export * from "./utils/PlayerCheck";
